refactor: rename utils.internal.ts to _utils.ts
• Move utils.internal.ts to _utils.ts for better file organization

• Update imports in all command files (daily, blocks, session, monthly)
